@@ -84,76 +63,82 @@
     }
   ],
   "dependencies": {
-    "symbol-observable": "^1.0.1"
+    "tslib": "^1.9.0"
   },
+  "deprecated": false,
   "description": "Reactive Extensions for modern JavaScript",
   "devDependencies": {
-    "@angular-devkit/build-optimizer": "0.0.24",
-    "babel-polyfill": "^6.23.0",
-    "benchmark": "^2.1.0",
+    "@angular-devkit/build-optimizer": "0.4.6",
+    "@angular-devkit/schematics": "^0.5.4",
+    "@types/chai": "4.1.2",
+    "@types/lodash": "4.14.102",
+    "@types/mocha": "2.2.48",
+    "@types/node": "9.4.5",
+    "@types/sinon": "4.1.3",
+    "@types/sinon-chai": "2.7.29",
+    "@types/source-map": "^0.5.2",
+    "babel-polyfill": "6.26.0",
+    "benchmark": "2.1.0",
     "benchpress": "2.0.0-beta.1",
-    "chai": "^3.5.0",
-    "color": "^0.11.1",
+    "chai": "4.1.2",
+    "color": "3.0.0",
     "colors": "1.1.2",
-    "commitizen": "^2.8.6",
-    "coveralls": "^2.11.13",
-    "cz-conventional-changelog": "^1.2.0",
-    "danger": "^1.1.0",
-    "doctoc": "^1.0.0",
-    "escape-string-regexp": "^1.0.5 ",
-    "esdoc": "^0.4.7",
-    "eslint": "^3.8.0",
-    "fs-extra": "^2.1.2",
-    "get-folder-size": "^1.0.0",
-    "glob": "^7.0.3",
-    "gm": "^1.22.0",
-    "google-closure-compiler-js": "^20170218.0.0",
-    "gzip-size": "^3.0.0",
-    "http-server": "^0.9.0",
-    "husky": "^0.13.3",
-    "klaw-sync": "^3.0.0",
-    "lint-staged": "3.2.5",
-    "lodash": "^4.15.0",
-    "madge": "^1.4.3",
-    "markdown-doctest": "^0.9.1",
-    "minimist": "^1.2.0",
-    "mkdirp": "^0.5.1",
-    "mocha": "^3.0.2",
+    "commitizen": "2.9.6",
+    "coveralls": "3.0.0",
+    "cross-env": "5.1.3",
+    "cz-conventional-changelog": "1.2.0",
+    "danger": "1.1.0",
+    "dependency-cruiser": "2.13.0",
+    "doctoc": "1.3.0",
+    "escape-string-regexp": "1.0.5",
+    "esdoc": "0.4.7",
+    "eslint": "4.17.0",
+    "fs-extra": "5.0.0",
+    "get-folder-size": "1.0.1",
+    "glob": "7.1.2",
+    "gm": "1.23.1",
+    "google-closure-compiler-js": "20170218.0.0",
+    "gzip-size": "4.1.0",
+    "http-server": "0.11.1",
+    "husky": "0.14.3",
+    "klaw-sync": "3.0.2",
+    "lint-staged": "7.1.1",
+    "lodash": "4.17.5",
+    "markdown-doctest": "0.9.1",
+    "minimist": "1.2.0",
+    "mkdirp": "0.5.1",
+    "mocha": "5.0.0",
     "mocha-in-sauce": "0.0.1",
-    "npm-run-all": "^4.0.2",
-    "npm-scripts-info": "^0.3.4",
-    "nyc": "^10.2.0",
-    "opn-cli": "^3.1.0",
-    "platform": "^1.3.1",
-    "promise": "^7.1.1",
-    "protractor": "^3.1.1",
+    "npm-run-all": "4.1.2",
+    "nyc": "11.4.1",
+    "opn-cli": "3.1.0",
+    "platform": "1.3.5",
+    "promise": "8.0.1",
+    "protractor": "3.1.1",
     "rollup": "0.36.3",
-    "rollup-plugin-inject": "^2.0.0",
-    "rollup-plugin-node-resolve": "^2.0.0",
+    "rollup-plugin-inject": "2.0.0",
+    "rollup-plugin-node-resolve": "2.0.0",
     "rx": "latest",
-    "rxjs": "latest",
-    "shx": "^0.2.2",
-    "sinon": "^2.1.0",
-    "sinon-chai": "^2.9.0",
-    "source-map-support": "^0.4.0",
-    "tslib": "^1.5.0",
-    "tslint": "^4.4.2",
-    "typescript": "~2.0.6",
-    "typings": "^2.0.0",
-    "validate-commit-msg": "^2.14.0",
-    "watch": "^1.0.1",
-    "webpack": "^1.13.1",
+    "rxjs": "^5.5.7",
+    "shx": "0.2.2",
+    "sinon": "4.3.0",
+    "sinon-chai": "2.14.0",
+    "source-map-support": "0.5.3",
+    "symbol-observable": "1.0.1",
+    "systemjs": "^0.21.0",
+    "ts-node": "6.1.0",
+    "tsconfig-paths": "3.2.0",
+    "tslint": "5.9.1",
+    "tslint-no-unused-expression-chai": "0.0.3",
+    "typescript": "~2.8.1",
+    "validate-commit-msg": "2.14.0",
+    "webpack": "1.13.1",
     "xmlhttprequest": "1.8.0"
-  },
-  "directories": {},
-  "dist": {
-    "integrity": "sha512-oRYoIKWBU3Ic37fLA5VJu31VqQO4bWubRntcHSJ+cwaDQBwdnZ9x4zmhJfm/nFQ2E82/I4loSioHnACamrKGgA==",
-    "shasum": "28d403f0071121967f18ad665563255d54236ac3",
-    "tarball": "https://registry.npmjs.org/rxjs/-/rxjs-5.5.2.tgz"
   },
   "engines": {
     "npm": ">=2.0.0"
   },
+  "es2015": "./_esm2015/index.js",
   "homepage": "https://github.com/ReactiveX/RxJS",
   "keywords": [
     "Rx",
