@@ -4,7 +4,7 @@ import { TimeContract } from '../contracts/time.contract';
 import { TimeContractStub } from '../contracts/time.contract.stub';
 import { ActivatedRoute } from '@angular/router';
 import { FormsModule } from '@angular/forms';
-import { Observable } from 'rxjs/Observable';
+import { Observable } from 'rxjs';
 
 describe('ClockComponent (shallow)', () => {
     let fixture: ComponentFixture<ClockComponent>;
