@@ -1,8 +1,8 @@
 import { Injectable } from '@angular/core';
-import { Observable } from 'rxjs/Observable';
+import { Observable } from 'rxjs';
 import { Configuration } from '../../configuration';
 import { Http, Headers, RequestOptions, Response } from '@angular/http';
-import { HttpClient } from '@angular/common/http';
+import { HttpClient } from '@angular/common/http'
 import { President } from '../models/president.model';
 import { PresidentialContract } from '../contracts/presidential.contract';
 @Injectable()
@@ -13,6 +13,7 @@ export class PresidentialService implements PresidentialContract {
         let route: string = this.configuration.routeToApi + "api/president";
         return this.httpClient.get<Array<President>>(route,{});
     }
+
 
     setPresidents(presidents:Array<President>):Observable<any>{
         let route: string = this.configuration.routeToApi + "api/president";
